style(web-builder): move config to shared/config to pass lint check
