+++ packages/builder/web-builder/src/core/createBuilder.ts
@@ -1,9 +1,8 @@
-import { pick } from '../shared';
+import { pick } from '../shared/utils';
 import { createContext, createPublicContext } from './createContext';
 import { createPluginStore } from './createPluginStore';
 import { initConfigs } from './initConfigs';
 import type { PluginStore, BuilderOptions } from '../types';
-import { PluginCss } from '../plugins/css';
 
 export function mergeBuilderOptions(options?: BuilderOptions) {
   const DEFAULT_OPTIONS: Required<BuilderOptions> = {
@@ -74,6 +73,7 @@ async function addDefaultPlugins(pluginStore: PluginStore) {
   const { PluginBabel } = await import('../plugins/babel');
   const { PluginTsLoader } = await import('../plugins/tsLoader');
   const { PluginTsChecker } = await import('../plugins/tsChecker');
+  const { PluginCss } = await import('../plugins/css');
 
   pluginStore.addPlugins([
     // Plugins that provide basic webpack config
