+++ packages/builder/web-builder/src/shared/config/normalize/index.ts
@@ -1,4 +1,4 @@
-import type { BuilderConfig, BuilderFinalConfig } from '../../types';
+import type { BuilderConfig, BuilderFinalConfig } from '../../../types';
 import normalizeSourceConfig from './source';
 
 export const normalizeConfig = async (
