+++ packages/builder/web-builder/src/types/index.ts
@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/no-restricted-imports */
 export * from './hooks';
 export * from './config';
 export * from './plugin';
 export * from './builder';
 export * from './context';
 export * from './thirdParty';
+/* eslint-enable */
